# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For DH Mobile ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vorchers/restore/{id}": {
            "put": {
                "tags": [
                    "Vorcher"
                ],
                "summary": "Khôi phục một vorcher đã bị xóa mềm",
                "operationId": "5fc52f158050ce90f4072ffedb253d35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Khôi phục vorcher thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
        "/api/vorchers/force-delete/{id}": {
            "delete": {
                "tags": [
                    "Vorcher"
                ],
                "summary": "Xóa vĩnh viễn vorcher đã xóa mềm",
                "operationId": "35cef9c013ea5b096d32c494268f19d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa vĩnh viễn thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}